feat(multiselect): prepare for useComboBox hook
