+++ docs/pages/multiselect.mdx
@@ -2,81 +2,7 @@
 
 ```jsx
 function MultiSelectExample() {
-  const fruits = [
-    "Acai",
-    "Apples",
-    "Apricots",
-    "Avocado",
-    "Ackee",
-    "Bananas",
-    "Bilberries",
-    "Blueberries",
-    "Blackberries",
-    "Boysenberries",
-    "Bread fruit",
-    "Cantaloupes(cantalope)",
-    "Chocolate - Fruit",
-    "Cherimoya",
-    "Cherries",
-    "Cranberries",
-    "Cucumbers",
-    "Currants",
-    "Dates",
-    "Durian",
-    "Eggplant",
-    "Elderberries",
-    "Figs",
-    "Gooseberries",
-    "Grapes",
-    "Grapefruit",
-    "Guava",
-    "Honeydew melons",
-    "Horned melon(Kiwano)",
-    "Huckleberries",
-    "Ita Palm",
-    "Jujubes",
-    "Kiwis",
-    "Kumquat",
-    "Lemons",
-    "Limes",
-    "Lychees",
-    "Mangos",
-    "Mangosteen",
-    "Mulberries",
-    "Muskmelon",
-    "Nectarines",
-    "Ogden melons",
-    "Olives",
-    "Oranges",
-    "Papaya",
-    "Passion fruit",
-    "Peaches",
-    "Pears",
-    "Peppers",
-    "Persimmon",
-    "Pineapple",
-    "Plums",
-    "Pluot",
-    "Pomegranate",
-    "Prickly Pear",
-    "Quince",
-    "Rambuton",
-    "Raspberries",
-    "Rose Apple",
-    "Starfruit",
-    "Sapadilla",
-    "Strawberries",
-    "Tamarind",
-    "Tangelo",
-    "Tangerines",
-    "Tomatoes",
-    "Ugli fruit",
-    "Voavanga(Spanish Tamarind)",
-    "Watermelons",
-    "Xigua melon",
-    "Yellow watermelon",
-    "Zucchini",
-  ];
+  const fruits = ["Acai", "Apples", "Apricots", "Avocado", "Ackee", "Bananas"];
 
   const options = fruits.map((d, i) => ({
     value: d,
@@ -88,7 +14,7 @@ function MultiSelectExample() {
 }
 ```
 
-## Searchable Multi Select
+<!-- ## Searchable Multi Select
 
 ```jsx
 function MultiSelectExample() {
@@ -178,9 +104,9 @@ function MultiSelectExample() {
     <MultiSelect isMulti options={options} placeholder="Select a fruit..." />
   );
 }
-```
+``` -->
 
-## Render Custom Placeholder
+<!-- ## Render Custom Placeholder
 
 ```jsx
 function MultiSelectExample() {
@@ -695,4 +621,4 @@ function MultiSelectExample() {
     />
   );
 }
-```
+``` -->
